Refactor middleware and enhance package scripts for environment management
- Removed the `config.ts` and `index.ts` middleware files, which contained Basic authentication logic and environment configuration.
- Updated `package.json` to include new scripts for staging and production environments, allowing for better management of application states during development and deployment.
