+++ src/middleware.ts
@@ -1,10 +1,48 @@
 import { NextResponse } from 'next/server';
 import type { NextRequest } from 'next/server';
-import { isBasicAuthEnabled, getBasicAuthCredentials, getEnvironment, config } from './config';
 
-export { config };
+// Middleware設定
+export const config = {
+  matcher: ['/:path*']
+};
+
+// 定数
+const PRODUCTION_DOMAIN = 'optiverse-now.com';
+const STAGING_DOMAIN = 'dev.optiverse-now.com';
+const LOCAL_DOMAIN = 'localhost:3000';
+
+// 環境判定用の関数
+const getEnvironment = (hostname: string | null): 'production' | 'staging' | 'local' | 'unknown' => {
+  if (!hostname) return 'unknown';
+  
+  if (hostname.includes(PRODUCTION_DOMAIN)) return 'production';
+  if (hostname.includes(STAGING_DOMAIN)) return 'staging';
+  if (hostname.includes(LOCAL_DOMAIN)) return 'local';
+  
+  return 'unknown';
+};
+
+// Basic認証の有効/無効判定
+const isBasicAuthEnabled = (): boolean => {
+  const appEnv = process.env.APP_ENV || 'develop';
+  console.log('Current APP_ENV:', appEnv);
+  return process.env.BASIC_AUTH_ENABLED === 'true';
+};
+
+// Basic認証の認証情報取得
+const getBasicAuthCredentials = () => {
+  const user = process.env.BASIC_AUTH_USER;
+  const password = process.env.BASIC_AUTH_PASSWORD;
+
+  if (!user || !password) {
+    throw new Error('Basic認証の環境変数が設定されていません');
+  }
+
+  return { user, password };
+};
 
 export function middleware(request: NextRequest) {
+  
   const hostname = request.headers.get('host');
   const environment = getEnvironment(hostname);
